Phase 0-2: HEAT_TERM rename, dual-write bridge, MLSAG wiring, BP+ stub, v12 fork

+++ src/SwapDaemon/PoolAMM.h
@@ -43,7 +43,7 @@ uint64_t poolGetInputAmount(uint64_t outputAmount,
 
 // Calculate fee amount from input.
 inline uint64_t poolGetFeeAmount(uint64_t inputAmount, uint32_t feeBps) {
-  return (inputAmount * feeBps) / 10000;
+  return (uint64_t)((unsigned __int128)inputAmount * feeBps / 10000);
 }
 
 // ─── LP share calculations ───────────────────────────────────────────
